Rearrange F5 'debug' lines for convenient use in normal game play
During normal game play some of the 'debug' output is actually often
indispensable:
- current coordinates (Where are my friends? Where are my bones?)
- current direction
- node pointed at (What's this funny plant? Is that silver, tin or mithril?)

But to see this you have to enable 3 lines of stuff most of which a normal
player doesn't care about (e.g. map seed). Together with a busy chat it can
make quite a mess of text.

Solution: the 3 essential things listed above easily fit in 1 (one!) line,
together with the MT version (show it in screenshots and YouTube videos).

The compass direction knows about NW, NE, SW, SE now, the angle is rounded
to full integers. The X/Y/Z coordinates, too, unless show_debug is enabled.
The reason for that: if you build a house and make one wall up to X=10.4
and the other up to X=10.6 your walls will have different lenghts, because
you should have rounded 10.6 to 11 in your head. Now MT does this for you.

The rest is mostly performance related (FPS, drawtime, dtime_jitter, RTT)
and fits all in the 2nd line, which is now only enabled if F5 is pressed
one more time. The map seed isn't performance, but I put it in line 2
anyway. To accomodate all I merged range_all into view_range (and dropped
the decimal place). The selected node's param2 stayed with the node type,
but is only shown if line 2 is visible, normal players don't need it.

To address @paramat's concerns, if the screen width appears too small the
MT version, map seed and pointed node's param2 (i.e. the non performance-
related items) get 'wrapped' to an additional middle line, enlarging the
debug info from 2 to 3 lines again on smaller screens. (Of course the chat
moves to follow suit.)

If only the status is enabled (for normal players) then it's still only 1
single line, which is shortened by hiding the MT version string.