Online menu: Do not show ping to server list, use `lag` instead.
The `ping` value is an entirely nonsense indicator, which gives players
an entirely wrong indication about game performance and in-game lag:

1. Since it indicates the value from the game server to the
servers.minetest.net serverlist, it bears zero correlation to the
actual ping of the player to the server list.

2. Because that relationship is absent, the player could actually
get a better ping than the serverlist server, since, a US player
connecting to a US server would get a much lower ping than indicated.

3. A player thinking that a low ping server would be good, could be
choosing a server that is high ping for them, since a player in Japan
or Australia isn't actually going to ever get a good ping value to
a server in Europe.

Fortunately, we already have a good number that indicates the game
server performance, it's just not based on network latency, but based
on `max_lag` - the indicator as to how well the server is capable of
handling game ticks in time.

As our default servertick is 0.09, and was 0.10, we should expect
most idle servers to be a few ms above 0.100. Busy servers will peak
out above 0.2.

I've studied the server list and made several spreads of the max_lag
values reported by each server. It turns out that of the 200 or so
servers reporting, 90% is under 0.125, and this indicates that they're
idle. We label these as "green".

We see only 10 or so servers that have loads above ~2x that value,
so I've chosen .275 as "red" indicating high load. At that value,
players are experiencing 100ms+ delays in all their actions, so it's
a good indication that the server isn't capable of keeping up or is
seeing "lag" spikes.

I've divided up the remaning gap between .125 and .275 in two equal
sections for "yellow" and "orange".

I believe this measure is much more fair than ping values. Yes it can
possibly be cheated by a server operator, but so can ping values,
because one can just choose a cloud host that is in the NL and
completely lie to the userbase. A amount of english spoken servers
are likely focussed more on non-UK player base, and are needlessly
penalized with "high ping" icons, and the same for many of the non-EU
local servers that target local player bases.