Fix LMB clicks on keybind button not being caught
In the GUI key binding change screen, after pressing the button
associated with an action, the GUI waits for a new key binding to be
input by the user. The next user input event (key or mouse press) should
be set as the new binding. However, this was not the case for LMB clicks
because the GUI was intercepting these clicks and treating them as a
different kind of event.

This commit adds a check that ensures LMB clicks on the action's button
are recognised.