Include dug nodes.
Mods that put metadata on nodes can preserve that metadata when the node is dug by defining an appropriate on_dig.
But this change makes that largely  unnecessary.  If the nodes drops itself when dug, and drops nothing else,
the metadata fields will be preserved in the item placed into the players inventory (or dropped into the world if the player is full).  The fields "description", "infotext" and "formspec" are not copied, since these have specific defined
functions.

This has been tested in 0.4.16 and 0.5.0-dev.