Preserve on falling only if a callback is defined.
To avoid calling this code all the time, the previous version had three restrictions:
- the node must have metadata
- the node must drop itself
- the field must not be one of the predefined fields

This version has a different way to avoid calling the code too often:  the node must request to be called.
This is done by a implementing a callback.  The callback is currently called "preserve_metadata", but that is certainly open for discussion.

Because the node has requested the callback, the old restrictions are no longer checked.  If a node has no metadata, but wants to add item metadata, perhaps a "description" field, it can do so.  If the node drops something other than itself, it could still put metadata on its drop.  If a node breaks in half when it drops, it could even put its metadata on one half but not the other (e.g. a grapevine could drop a bare vine and also grapes with metadata indicating how ripe or how sweet they are - wine:malaga_dulce anyone?)