Normalmaps: Apply only specific texture modifiers
Texture modifiers such as colorize must not be applied on normalmaps, however something like transform (e.g. rotating) should be applied.
Additionally texture modifiers are not longer attempted to be applied on missing normalmap images, this fixes error messages.
The error log regarding a missing result image is also disabled for normalmaps because normalmaps are optional.