Digging and placing: Remove redundant mapblock sending
When placing, minetest.set_node (or something similar) is executed, so the changed nodes are sent via a TOCLIENT_ADDNODE.
Therefore, the mapblock does not need to be sent back to the client.
However, if a node is protected, placement fails and the client has to be informed that its placement prediction is wrong.
Digging is changed similarly.