Use TOCLIENT_ADDNODE instead of mapblock sending when possible
Also, clean up code a bit, e.g. peer_id is used very often.

When a node contains metadata, the mapblock needs to be sent because the client removed it when predicting a node.
The whole mapblock is still sent for edge cases like interacting too far or when being dead.