Fix player's attachments not updating visibility correctly
I found and solved this issue after noticing that several mods I have made would not display items attached to the player properly.

More specifically, when an item is attached to a player while they are in third person, it will not be visible to them until they refresh their view. This is a solution to that problem. 

On another note, I would also suggest implementing some option (for example, "visible_in_first_person") for players' children to be shown to them even in first person, as that would make things easier for my hangglider mod.