Optimize map sector unload
Copying of v2s16 values in a loop control value in
deleteSectors scanning loop slows the method down,
introducing lag when there are many loaded sectors to go
through when looking for some to delete. Convert the
offending loop control variable into a reference to
avoid the unnecessary copy as the only thing the value
is used for is indexing the array of loaded sectors.