Optimize active mapblock updating
There are three loops which have a v3s16 control variable but
the only thing they do with it is to move these values around
or index the table of active blocks to get the blocks to be
updated. Convert them to constant references to avoid the
slowdown from the unnecessary copying of these values.