Optimize clearObjects
This method has two loops with v3s16 control variables and
the only thing its value is used for is to get the active
mapblocks so they can be pinned down before the objects are
enumerated and removed and unpinned after this job is done.
Convert these control variables to const references to
avoid the unnecessary copying.