Optimize cave noise parameter constructors
These constructors are getting the "chunksize" parameter by
value but the only thing they do with it is to store the
thing in a field. This means the value is copied twice per
constructor call, slowing the mapgen down. Turn the parameter
into a const reference to get rid of the unnecessary
parameter passing copy, leaving only the final copy in place.