Optimize caves carving
Each time a cave needs to be carved, the vector which
specifies the direction is passed by value but the only thing
it is used for is to calculate the "real" vector used in the
actual cave path generator by scaling it and stuff. Eliminate
the unneeded value passing to speed this thing up.