Fix debug.txt growing bigger and bigger
Before opening the file for writing, its file size is tested. If it exceeds 1 MB, it is deleted, otherwise the log is appended to the old messages.
Additionally, information about the log file location and whether it is deleted is printed to the action log when starting minetest.

The function which opens the file is now called setFile because "open" sounds too simple for what it does.