Push exceptions out of Map::getNodeNoCreate
Delete Map::getNodeNoCreate which throws an exception, rename
Map::getNodeNoCreateNoEx to take its place and split
Map::getNode into Map::getNodeNoEx (which does all the logic
and returns true/false) and Map::getNode (which calls
Map::getNodeNoEx under the hood and throws the exception on
false).

This gets rid of the pesky "NoEx" suffix in most of the code
and prepares the way for complete removal of the exception
in Map::getNode.