Rename MapSector::getBlockNoCreateNoEx
Mapsector has MapSector::getBlockNoCreateNoEx but no
MapSector::getBlockNoCreate, probably because exception
throwing was killed in that method long ago. Get rid of
the "NoEx" in that method name as well now as there will
be no need to stress that fact about no exception usage
once this series is completed.