(#3075) Hackish but simple fix for the ~5yo waving water bug
Resolves #3075

1. Extract the code that determines the tile index by the face
direction so it can be reused in other places. Specifically in the
`face_contents` function which determines which face of the adjacent
nodes to hide based on their real and visual solidness.

2. Extend `face_contents` to accept tile index which determines the
direction of the face and force showing the top liquid face when the
waving water is enabled and the face has a more solid block on top of
it.

Not the best solution but it kinda works. Please let me know what you
think about it. Also any ideas on how to improve the solution are
really appreciated.