Fix Debian build dependencies in README
This patch replaces the statement in the README about a build dependency of the
minetest binary on the package “build-essential” with a corresponding statement
about build dependencies on the three packages “g++”, “make”, and “libc6-dev”.

Though often used as a shorthand, “build-essential” is not a package used to
install packages that are essential for building software – “build-essential”
depends on packages that are essential for building Debian packages. Therefore,
the dependency on “build-essential” implies a dependency on a lot of packages that
are not actually necessary to build minetest (e.g. the Perl programming language).