Add ObjectRef:has_priv(priv_name) Lua call
Most of our calls are very generic and involved loading all the privileges from the Lua stack to the core then process & dump again to Lua.
With this call we offer a very short way to check if a ObjectRef loaded in game has a privilege directly.

This prevent the common way: minetest.check_player_privs(player:get_player_name(), [interact]) which involved many callbacks to and from the core
Now just call player:has_priv(interact) and it works (it's not very common to check multiple privs at the same time too)