Added serialization routines for double precision floating point.
Added routines to read and write D1000 format, which reads/writes 1000. * x as a 32-bit signed integer.
Added rotines to read and write D format, which represents a double as the sum of two f32 values.
Added routines to read and write v2d in V2D format.
Added routines to read and write v32 in V3D1000 and V3D formats.