merge Thread::kill() inside Thread::~Thread()
The Thread kill() method isn't used outside testThreadKill unit test
and Thread destructor. It could be make private.

The testThreadKill test should be removed: it tests the kill() feature
in a way minetest doesn't use it (kill and restart a thread), specially
because some pthreads implementations on BSD will not allow to lock
again the m_start_finished_mutex which is hold when the thread is
killed.

Closes: #6065