move build/android to root dir
The top level `build` folder is oftentimes the default build folder for
CMake. Move the `build/android` folder to the root of the project, which
is the only thing in `build` folder. Now the `build` folder can be used
for CMake.

Update the Android project files, because the `android` dir is now one
level less deep relative to the root of the project (effectively
replacing `../../..` with `../..`

Also update the `util/bump_version.sh` script expecting `build/android`
folder.

Ignore the possible `build` folder (add to `.gitignore`)

Fixes: https://github.com/minetest/minetest/issues/9800