correct media downloader initialization
auto incorrectly determined the type of the media downloader pointer to be a raw pointer instead of a unique pointer. This caused a type mismatch in the std::pair constructor - it expected a unique pointer - and fails on GCC 4.9. It seems it should also fail on newer versions, unless this was a compiler bug in GCC 4.9. Explicitly initializing the type to a unique pointer, and using move semantics to transfer ownership to the pair constructor, is the correct way to handle it, regardless of the error.