use shared pointer for pending media
Using a unique pointer and keeping a raw pointer around is risky. Changing a unique pointer type to a shared pointer type without careful thought is also risky. Fortunately, in this case it's easy. The pending media is only used in one method in the client, and one in the client packet handler, and the contents are not copied in either case. No unecessary copies of the shared pointer will be sitting around, and everything will clean itself up in an orderly, and safe fashion.