Increase the default FPS cap to 200, including on Android
This effectively uncaps the framerate on all platforms, making the
game run as fast as possible for smoother and lower-latency gameplay.
On platforms that force V-Sync such as Android, the framerate will
not exceed the display refresh rate. (`vsync` can be set to `true`
in `minetest.conf` on other platforms.)

On desktop platforms, this allows making use of high refresh rate
monitors without requiring any configuration from the user. 120 Hz
and higher monitors are becoming common, even on some non-gaming
laptops as of late.

On fast Android devices, this provides a much better out of the box
experience by allowing the game to run at 60 FPS instead of 30 FPS.

The value 200 was chosen because it's enough for most monitors
on the market, while also reducing audible coil whine on high-end GPUs
and avoiding physics glitches due to high FPS.