Correct "Node Timers" specification
I believe that the specification is incorrect, and that the node timers appear at the end of a mapblock for map format version >= 25, not just map format version 25.  This is based on my observations of map block format #28 (from minetest-5.4.x and multicrraft-2.0.x) while developing a stand-alone C++ library for reading and writing minetest mapblock data.