Do not stat /etc/localtime all the time
tzset(3) initializes the time zone from the $TZ environment variable,
or, if the $TZ variable is not set, from /etc/localtime.  According to
POSIX.1-2001, localtime(3) is required to behave as though tzset(3) was
called, while localtime_r(3) is not. This translated into many syscalls
to check /etc/localtime as every log line & mapblock needs a timestamp.

The fix uses localtime_s() on Windows, as it has no localtime_r(). Both
functions are threadsafe – localtime() is not, due to the static struct.