Fix undefined behavior in TileLayer::operator==
TileLayer::operator== compares the `color` and `scale` fields,
but these may be uninitialized. `color` should only be accessed
if `has_color` is true. `scale` should be initialized with a
sensible default.

This _usually_ would not cause any issues, but since it is
undefined behavior, the compiler would be free to do anything,
which could result in hard-to-debug problems.