Fix texture string overlay with texture groups of different sizes
See #12459.

Texture overlays of the form `cobble.png^(external_texture.png)` break
if it is not guaranteed that both texture parts have the same size.
Overlaying a texture group previously did not check for different sizes.

Because older mods might rely on the undocumented behavior,
overlaying a texture group in single parentheses continues to not check sizes,
and only texture groups in double parentheses are resized.

(A texture group is a part of a texture string enclosed in parentheses.)