Add setting to force enable HRTF
Head-related Transfer Function is an OpenAL setting that improves
the quality of 3d sounds when wearing headphones. Rather than just
changing stereo balance/volume, it also changes the time offset.
This results in sounds being much clearer and easier to locate.

OpenAL will automatically enable HRTF when it detects headphones.
Unfortunately, it's unable to do this on Windows, pipewire, or
ALSA, and it's unreliable for pulseaudio.